Préparation tower next season (#234)
* refactor: tower season and message
