+++ slash_commands/subcommands/tower/info-boss.js
@@ -1,6 +1,6 @@
-const { displayHealth, getRandomPrivateJokes } = require("./valider-jeu");
 const { TowerBoss, GuildConfig } = require("../../../models");
 const { EmbedBuilder } = require("discord.js");
+const { displayHealth, getRandomPrivateJokes } = require("../../../util/events/tower/towerUtils");
 
 const infoBoss = async (interaction, options) => {
     const guildId = interaction.guildId;
